Reorganize files in rich text editor
