+++ lib/widgets/rich_text_editor/src/nodes/quote_node.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:junto_beta_mobile/widgets/rich_text_editor/rich_text_editor.dart';
+
+import 'rich_text_node.dart';
 
 class QuoteNode extends StatelessWidget {
   const QuoteNode({
