+++ lib/widgets/rich_text_editor/src/nodes/text_node.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:junto_beta_mobile/widgets/rich_text_editor/rich_text_editor.dart';
 
+import 'rich_text_node.dart';
 import 'types.dart';
 
 class TextNode extends StatelessWidget {
