+++ lib/widgets/rich_text_editor/src/selectable_inline_image.dart
