+++ lib/widgets/rich_text_editor/src/text_span_controller.dart
